pyproject.toml
[tool.poetry]
name = "netdevconfigurator"
version = "0.1.0"
description = ""
authors = ["Shmonov Matvey Vasilievich <meology77@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
ttkwidgets = "^0.13.0"
psycopg2-binary = "^2.9.10"
pyyaml = "^6.0.2"
pyserial = "^3.5"
paramiko = "^3.5.0"
scapy = "^2.6.1"
wrapt-timeout-decorator = "^1.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
debugpy = "^1.8.13"
alembic = "1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
